Testplan

+++ src/main/java/com/shopify/dto/InventoryDTO.java
@@ -1,5 +1,12 @@
 package com.shopify.dto;
 
+/**
+ * Interface that represents a data transfer object which is mapped to queries
+ * from database tables.
+ * 
+ * @author stephen
+ *
+ */
 public interface InventoryDTO {
 	long getId();
 
@@ -7,5 +14,5 @@ public interface InventoryDTO {
 
 	int getQuantity();
 
-	int getPrice();
+	double getPrice();
 }
